requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
constantly==15.1.0
cryptography==39.0.0
daphne==3.0.2
Django==4.1.6
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
Pillow==9.4.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
python-decouple==3.7
pytz==2022.7.1
redis==4.4.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
zope.interface==5.5.2
